refactor(blockchain): type SNAKE_CHAIN_LENGTH as u32 + test Unstorable arm (Story 4.3 review)
Follow-up from the Story 4.3 bmad-code-review re-run (2 low-severity patch
findings, no correctness defects):

- SNAKE_CHAIN_LENGTH is a window *width* compared against u32 block sequences,
  never an array length, so declare the const generic as `u32` instead of the
  array-sizing `usize`. Removes the unguarded `SNAKE_CHAIN_LENGTH as u32`
  narrowing cast in classify_block's FR60 gate at its root, rather than guarding
  it — the truncation the reviewers flagged is now impossible by type. (The u64
  widening of `s_head + w` inside snake_chain_window_verdict stays: that guards
  the u32 *sum*, a distinct concern.)
- Add receive_block_operational_refusal_is_unstorable: exercises the previously
  untested AdmitError::{TableFull,StorageSaveFailed} -> Rejected(Unstorable)
  mapping (storage seam binds first in the test fixture; asserts Unstorable, not
  an InvalidEvidence validity verdict, and that the block is not stored).

78 tests pass (was 77); clippy unchanged (2 pre-existing warnings); release
build green; no dep drift. Diff scoped to api.rs + intake.rs.

+++ src/api.rs
@@ -207,7 +207,7 @@ pub struct Blockchain<
     Storage: StorageTrait,
     Config: ChainConfigTrait,
     const MAX_NODES: usize,
-    const SNAKE_CHAIN_LENGTH: usize,
+    const SNAKE_CHAIN_LENGTH: u32,
     const VERIFICATION_HORIZON: usize,
     const MAX_BLOCKS: usize,
     const MAX_BRANCH_COUNT: usize,
@@ -238,7 +238,10 @@ pub struct Blockchain<
 
     // Real snake-chain state is two block-table indices, not a W-sized
     // window. `SNAKE_CHAIN_LENGTH` remains an algorithmic bound for
-    // maintaining the tail index relative to the active head.
+    // maintaining the tail index relative to the active head. It is a `u32`
+    // (not the array-sizing `usize` the other const generics use) because it
+    // is a window *width* compared directly against `u32` block sequences —
+    // never an array length — so it needs no cast at the FR60 comparison site.
     _active_chain_head_idx: u32,
     _snake_chain_tail_idx: u32,
     //
@@ -257,7 +260,7 @@ impl<
     Storage: StorageTrait,
     Config: ChainConfigTrait,
     const MAX_NODES: usize,
-    const SNAKE_CHAIN_LENGTH: usize,
+    const SNAKE_CHAIN_LENGTH: u32,
     const VERIFICATION_HORIZON: usize,
     const MAX_BLOCKS: usize,
     const MAX_BRANCH_COUNT: usize,
